requirements.txt
Flask==3.0.0
pandas==2.1.0
biopython==1.81
